Add `useRedirectToMain` hook
